Extract common toggle component
